separate validators, import print properly, clean up
